backend/requirements.txt
openai==2.15.0
httpx==0.28.1
python-dotenv==1.2.1
fastapi==0.128.0
SQLAlchemy==2.0.45
alembic==1.18.0
uvicorn==0.40.0
pydantic==2.12.5
httpx-socks[asyncio]==0.10.0
playwright==1.49.1
playwright-stealth==1.0.6

